requirements.txt
amqp==5.0.6
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.3
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crispy-forms==1.11.2
django-currencies==0.10.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
docker==5.0.2
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonfield==2.1.1
kombu==5.1.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
pywin32==227
rabbitmq==0.2.0
redis==3.5.3
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
